requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.10.0
Flask==2.2.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-whois==0.8.0
requests==2.28.2
requests-file==1.5.1
six==1.16.0
soupsieve==2.4
tldextract==3.4.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0